+++ dbaas/account/models.py
@@ -95,6 +95,21 @@ class Team(BaseModel):
 
         return list(all_users.difference(Set(users)))
 
+    def databases_in_use_for(self, environment):
+        from physical.models import DatabaseInfra
+        from logical.models import Database
+
+        infras = DatabaseInfra.objects.filter(environment=environment)
+        dbs = Database.objects.filter(team=self, databaseinfra__in=[infra.id for infra in infras])
+        
+        return dbs
+
+    def count_databases_in_use(self, environment):
+        try:
+            return len(self.databases_in_use_for(environment))
+        except Exception, e:
+            LOG.warning("could not count databases in use for team %s, reason: %s" % (self, e))
+            return 0
 
 def sync_ldap_groups_with_user(user=None):
     """