+++ dbaas/dbaas/settings.py
@@ -185,7 +185,30 @@ INSTALLED_APPS = (
     'django.contrib.admindocs',
     'raven.contrib.django.raven_compat',
     'ganalytics',
+    'haystack',
 )
+
+
+
+TEMPLATE_CONTEXT_PROCESSORS = (
+    "django.contrib.auth.context_processors.auth",
+    "django.core.context_processors.debug",
+    "django.core.context_processors.i18n",
+    "django.core.context_processors.media",
+    "django.core.context_processors.static",
+    "django.core.context_processors.tz",
+    "django.contrib.messages.context_processors.messages",
+    'django.core.context_processors.request',
+)
+
+# backend for haystack
+HAYSTACK_PATH =  os.getenv('HAYSTACK_PATH', os.path.join(SITE_ROOT, '../','whoosh_index'))
+HAYSTACK_CONNECTIONS = {
+    'default': {
+        'ENGINE': 'haystack.backends.whoosh_backend.WhooshEngine',
+        'PATH': HAYSTACK_PATH,
+    },
+}
 
 if not DB_ENGINE.endswith('sqlite3'):
     # support migrations