+++ dbaas/logical/forms/database.py
@@ -83,8 +83,7 @@ class DatabaseForm(models.ModelForm):
                 raise forms.ValidationError(_("Invalid plan for selected environment."))
 
             #validate if the team has available resources
-            infras_in_environment = DatabaseInfra.objects.filter(environment=environment)
-            dbs = Database.objects.filter(team=team, databaseinfra__in=[infra.id for infra in infras_in_environment])
+            dbs = team.databases_in_use_for(environment)
             database_alocation_limit = team.database_alocation_limit
             if (database_alocation_limit != 0 and len(dbs) > database_alocation_limit):
                 LOG.warning("The team %s has exceeded the database alocation limit of %s: %s" % (team,